Update SAS Engine MOD 2.8.7 western --- cleanup meaningless cast nests. New 360 degree Over Head Approach function for AI landing. AI cruise altitude / speed control rework. Add 10x engines aircraft. Debug AircraftHotkeys.class internal number duplication. RWR utility more historical functions. Multirole jet AI ground attack procedure expanded. SACLOS missile guiding Interface added.

+++ branches/dev_western/SAS_Engine_Mod_2.8_western/com/maddox/il2/objects/weapons/Rocket.java
@@ -152,8 +152,8 @@ public class Rocket extends ActorMesh implements MsgCollisionRequestListener, Ms
 		if (f3 > 0.0F) DM = (f1 - f2) / (f3 / Time.tickConstLenFs());
 		else DM = 0.0F;
 		P = f4;
-		timeFire = (long) ((double) (f3 * 1000F) + 0.5D);
-		timeLife = (long) ((double) (f5 * 1000F) + 0.5D);
+		timeFire = (long) (f3 * 1000F + 0.5F);
+		timeLife = (long) (f5 * 1000F + 0.5F);
 	}
 
 	public void start(float f) {
